Fixes: #13
