+++ algorithms/sort/HeapSort.java
@@ -1,3 +1,5 @@
+package Algorithms.Sort;
+
 /*********************************************************************
  * This code makes a Priority Queue and returns tha max value from that
  * so the nam MaxPriorityQueue.
